setup: add tauri and modular structure
